Blog section added to navbar and current blogs added to the list
This is far from being an ideal solution but jekyll primer we're using with this bootstrap doesn't let us you jekyll cards as are usually used (normally you'd put your blogs into a folder called _posts and then name them like this: YYYY-MM-DD-name.md but it wont compile for me when I try and google tells me its because github pages dont support jekyll plugins in some themes, including ours). For now the solution is that every new blog needs to be manually added to a blog section in data (same as any resource). Hopefully we can update the theme in the future and then we can do auto-generation of blog lists. This will do until then.